# docs/workflow_openapi.yaml
openapi: 3.0.3
info:
  title: Fast Lease API
  version: "1.0.0"
  description: >
    API для управления бизнес-процессом Fast Lease (Дубай):
    сделки, статусы, задания, вебхуки интеграций, метрики.
servers:
  - url: https://api.fastlease.local/v1
    description: Prod (пример)
  - url: https://staging.fastlease.local/v1
    description: Staging
...
paths:
...
  /webhooks/esign:
    post:
      tags: [Webhooks]
      summary: Вебхук провайдера e-sign (DocuSign/UAE Pass)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EsignEvent"
      responses:
        "204": { description: Принято }
  /webhooks/bank:
    post:
      tags: [Webhooks]
      summary: Вебхук банка о платеже
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankEvent"
      responses:
        "204": { description: Принято }
  /webhooks/aecb:
    post:
      tags: [Webhooks]
      summary: Вебхук AECB (кредитный отчёт/скор)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AecbEvent"
      responses:
        "204": { description: Принято }
...
components:
...
  schemas:
...
    EsignEvent:
      type: object
      required: [deal_id, status]
      properties:
        deal_id: { type: string, format: uuid }
        status: { type: string, enum: [COMPLETED, DECLINED] }
        envelope_id: { type: string }
        meta:
          type: object
          additionalProperties: true
    BankEvent:
      type: object
      required: [deal_id, kind, status]
      properties:
        deal_id: { type: string, format: uuid }
        kind: { type: string, enum: [ADVANCE, SUPPLIER] }
        status: { type: string, enum: [CONFIRMED, FAILED] }
        amount: { type: number, format: double }
        currency: { type: string }
        external_ref: { type: string }
    AecbEvent:
      type: object
      required: [deal_id, aecb_score, approved]
      properties:
        deal_id: { type: string, format: uuid }
        aecb_score: { type: integer }
        approved: { type: boolean }
        notes: { type: string }
        raw:
          type: object
          additionalProperties: true
...